fix(security): use int64 for compaction error to prevent integer overflow
Changes NewCompactionError signature from (uint, uint) to (int64, int64)
to match the actual types of database IDs (BIGSERIAL/int64).

This fixes an incorrect integer conversion warning from CodeQL that could
potentially cause:
- Negative value wrap-around (int64(-1) → uint(max))
- Overflow on 32-bit systems (int64 > uint32_max)
- Misleading error messages in edge cases

The fix simplifies the code by removing unsafe type conversions and
ensures semantic correctness since database auto-increment IDs are
signed 64-bit integers.

Impact:
- Only affects internal error handling (pkg/db package)
- No breaking changes to public API
- All tests pass (including TestCompactionError)

Resolves: https://github.com/jrmatherly/kinm/security/code-scanning/1

+++ pkg/db/db.go
@@ -224,7 +224,7 @@ func (d *db) list(ctx context.Context, namespace, name *string, rev int64, after
 	// is important for when the compaction ID is greater than any existing ID in the table. That can happen
 	// after a compaction where the last row was a delete=true row.
 	if rev != 0 && meta.ListID != 0 && meta.ListID < meta.CompactionID {
-		return meta, nil, errors.NewCompactionError(uint(meta.ListID), uint(meta.CompactionID))
+		return meta, nil, errors.NewCompactionError(meta.ListID, meta.CompactionID)
 	}
 
 	return meta, records, tx.Commit()
